backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# ML & Audio Processing
torch==2.1.2
torchaudio==2.1.2
openai-whisper==20231117
transformers==4.37.2
pyannote.audio==3.1.1
librosa==0.10.1
soundfile==0.12.1
spacy==3.7.2

# ONNX Optimization
onnx==1.15.0
onnxruntime==1.17.0
optimum[onnxruntime]==1.16.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1